Have done 1 homework from React
