+++ .eslintrc.json
@@ -22,6 +22,12 @@
         "no-alert": 0,
         "func-names": 0,
         "max-len": 0,
-        "linebreak-style": 0
+        "linebreak-style": 0,
+        "no-useless-concat": 0,
+        "import/prefer-default-export": 0,
+        "import/no-mutable-exports": 0,
+        "import/extensions": 0,
+        "prefer-const": 0,
+        "no-plusplus": 0
     }
 }
