+++ Triphouse-project/index.html
@@ -341,6 +341,6 @@
             </section>
         </footer> -->
     </div>
-<script src="./main.js"></script>
+<script type="module" src="./main.js"></script>
 </body>
 </html>
