+++ Triphouse-project/modules/filter-values.js
@@ -1,0 +1,5 @@
+export let filterValues = {
+  adults: 1,
+  children: 0,
+  rooms: 1,
+};
